Feature/m1 build (#3)
Problem

Building and running tests as outline in CONTRIBUTING.MD on a M1 based MacBook is not possible (or easy). This is due to various containers we rely on not having supporting the appropriate architecture. This results in strange crashes, hangs, etc when trying to build and test the project.

Solution

Update docker-compose and Docker files to address this. Included a README to describe how to run the m1 based docker-compose.

Checklist

@getquill/maintainers